requirements.txt
asgiref==3.3.4
autopep8==1.5.6
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
joblib==1.0.1
mysqlclient==2.0.3
numpy==1.20.3
pycodestyle==2.7.0
PyJWT==2.0.1
python-dotenv==0.17.0
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
